feat: board2board SPI example
You can enter the message by serial monitor and send it to the board by USART.
After that this code sends the message by SPI.

Slave well receive message from master.
However, when message is send from slave to master, 0 and 2 index characters are lost.

This bug should be fixed in the future.